Fixed wrong redirection in form placeAd, fixes #87, Made few values to doubles and added more logical number steps in forms, fixes #82, A premium user can edit the auction step price, fixes #32

+++ src/main/java/ch/unibe/ese/team1/controller/pojos/forms/SearchForm.java
@@ -32,7 +32,7 @@ public class SearchForm {
 	
 	@NotNull(message = "Requires a number")
 	@Min(value = 0, message = "In your dreams.")
-	private Integer prize;
+	private double prize;
 	
 	private boolean includeRunningCosts;
 
@@ -54,11 +54,11 @@ public class SearchForm {
 		this.radius = radius;
 	}
 
-	public Integer getPrize() {
+	public double getPrize() {
 		return prize;
 	}
 
-	public void setPrize(Integer prize) {
+	public void setPrize(double prize) {
 		this.prize = prize;
 	}
