+++ src/main/java/ch/unibe/ese/team1/controller/service/AlertService.java
@@ -124,7 +124,7 @@ public class AlertService {
 	 */
 	@Transactional
 	public void triggerAlerts(Ad ad) {
-		int adPrice = ad.getPrize();
+		double adPrice = ad.getPrize();
 		Iterable<Alert> alerts = alertDao.findByPriceGreaterThan(adPrice - 1);
 
 		// loop through all ads with matching city and price range, throw out
