+++ src/main/java/ch/unibe/ese/team1/model/dao/AdDao.java
@@ -9,15 +9,15 @@ import ch.unibe.ese.team1.model.User;
 public interface AdDao extends CrudRepository<Ad, Long> {
 	
 	/** this will be used if both houses AND flats are searched */
-	public Iterable<Ad> findByPrizeLessThan (int prize);
+	public Iterable<Ad> findByPrizeLessThan (double prize);
 
 	/** this will be used if only houses or flats are searched */
-	public Iterable<Ad> findByFlatAndPrizeLessThan(boolean flat, int i);
+	public Iterable<Ad> findByFlatAndPrizeLessThan(boolean flat, double i);
 
 	public Iterable<Ad> findByUser(User user);
 	
 	@Query("select ad from Ad ad where ad.prize + ad.runningCosts < ?1")
-	public Iterable<Ad> findByPrizeIncludingRunningCostsLessThan (int prize);
+	public Iterable<Ad> findByPrizeIncludingRunningCostsLessThan (double prize);
 	
 	@Query("select ad from Ad ad where ad.flat=?1 and ad.prize + ad.runningCosts < ?2")
 	public Iterable<Ad> findByFlatAndPrizeIncludingRunningCostsLessThan(boolean flat, int i);
